changed to new jhu file, fixed week naming
