+++ main.py
@@ -3,10 +3,8 @@ import process_inputs.cbsa_fips_mapping as cbsa
 import process_inputs.fips_place_cbsa as fpc
 import process_inputs.airports as air
 import process_inputs.build_nodes as bn
-import process_inputs.jhu_2 as jhu2
-#import process_inputs.jhu as jhu
+import process_inputs.jhu as jhu
 import process_inputs.adjacency_edges as adje
-import process_inputs.deaths_by_node as dbn
 
 from classes.node import Node
 from classes.fips import Fips
@@ -25,12 +23,12 @@ cbsa.build_CBSA_maps()
 # 3 - fips_place_cbsa
 # used for associating an airport to a county
 # returns a lookup for place_cbsa["place name"] -> [fips_code, cbsa_code]
-#place_cbsa = fpc.build_fips_maps()
+place_cbsa = fpc.build_fips_maps()
 
 # 4 - airports
 # link airports to CBSAs and FIPS
-#airports = air.process_airports(place_cbsa)
-#del place_cbsa # need to check if there are other references out there - this file is only used to process airports
+airports = air.process_airports(place_cbsa)
+del place_cbsa # need to check if there are other references out there - this file is only used to process airports
 
 # 5 - construct Nodes by CBSA for counties that have them, or by FIPS for those without
 bn.process_fips_for_nodes()
@@ -38,14 +36,11 @@ bn.process_fips_for_nodes()
 # 6 - process JHU daily files into weekly reports
 #jhu.create_weekly_reports('2022-01-01', '2022-03-31')
 #jhu.create_weekly_reports()
-jhu2.create_weekly_reports('2022-01-03', '2022-04-28')
+jhu.create_weekly_reports('2022-01-03', '2022-04-28')
 
 # 7 - process county adjacencies
-#adje.build_edges()
+adje.build_edges()
 
-# 8 - process deaths and attach to Node
-#dbn.count_deaths_by_node(True)
-#dbn.count_deaths_by_node()
 
 print("you are in main")
 breakpoint()
