+++ process_inputs/build_nodes.py
@@ -5,14 +5,16 @@ from classes.airport import Airport
 
 def process_fips_for_nodes():
     for [fips_code, fips] in Fips.collection.items():
+        node = None
+        cbsa = None
         if fips.cbsa_code == None:
-            node = Node("FIPS")
+            node = Node("FIPS", fips.county + _ + fips.state)
             node.airports = node.airports.union(fips.airports)
         else:
             cbsa = Cbsa.collection[fips.cbsa_code]
             node = Node.get_by_cbsa(cbsa.code)
             if not node:
-                node = Node("CBSA")
+                node = Node("CBSA", cbsa.name)
             node.airports = node.airports.union(cbsa.airports)
             # associate CBSA with UID
             cbsa.node_id = node.id_
